model service almost finished except for predict output being weird
