--- src/model_service/spam.py
@@ -1,4 +1,0 @@
-import lib_ml.foo
-
-def eggs():
-    return lib_ml.foo.bar()
